Limpieza de primavera 🌼
Refactorizada clase Facturae en múltiples archivos dentro del paquete josemmo\Facturae\Controller para:
 - Mejorar la legibilidad del código
 - Facilitar la implementación de nuevas funcionalidades
 - Permitir la creación de clases "a medida" que implementen los métodos estrictamente necesarios
